+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
  - Change web UI restart command
  - Fix issue saving Timer options ([#334](https://github.com/kizniche/mycodo/issues/334))
+ - Fix Output Usage error
 
 
 ## 5.4.9 (2017-11-27)